[ChaosVD] Improvements to how collision data can be inspected + performance and safety fixes.
- Implemented the Tollkit host interface to SChaosVDMainTab, so it can be passed to the Editor Mode Tools.
- Removed Code duplication in SChaosVDMainTab and added a way to get specific tabs if they are loaded.
- Added a way for CVD Tabs to generate an error widget that can be displayed if their setup failed.
- SChaosVDMainTab is now passed as a weak ptr instead of a raw ptr to any CVD tab.
- Removed collision data from the particle actors themselves and created a new Collision Data component per represented solver. This removes a lot of overhead caused by copying collision data to each particle, every frame. Also most that data was duplicated.
- Create a new SolverInfoActor. This is used by FChaosVDScene to centralize all the relevant info per represented solver.
- Re-Implemented collision data visualization as an UE Component Visualizer (and the ground work to be able to add more component visualizers if desired). Particle Data visualization has been kept with the CVD version of the visualizer (which is more limited but lighter).
- Added a way to debug draw (visualize) collision data without needing to have a particle actor selected.
- Created a new "Collision Data Inspector". As the collision data is not part of the CVD particle actor anymore, it cannot be inspected on its details panel. This new collision data inspector offers a dropdown list will all the collision data for the selected object.
- Added a way to click a visualized contact in the viewport and bring its details to the new Collision Data Inspector view.
- Added a way to see all the collision data for a particle in the new collision data inspector by clicking a button on the Particle's details panel.
- Downgraded two ensures to warning logs. There can happen during live debugging and it is not an issue.
- Updated SChaosVDNameListPicker so it auto-disables itself when the list is empty. Also it clearly states it is empty. This widget is used in a few places of the CVD UI.
- Disable navigation relevancy on components generated by CVD. The navigation system is already disabled for the CVD world, but the components having navigation relevancy enabled by default was having an impact because of the checks done when the actors are moved.

#jira UE-193750 UE-189695 UE-192481
#rb benn.gallagher

[CL 27677839 by sergio gardeazabal in ue5-main branch]

+++ Engine/Plugins/Experimental/ChaosVD/Source/ChaosVD/Public/ChaosVDEditorSettings.h
@@ -7,6 +7,8 @@
 #include "Engine/StaticMesh.h"
 #include "UObject/Object.h"
 #include "UObject/SoftObjectPtr.h"
+#include "Visualizers/ChaosVDParticleDataVisualizer.h"
+#include "Visualizers/ChaosVDSolverCollisionDataComponentVisualizer.h"
 
 #include "ChaosVDEditorSettings.generated.h"
 
@@ -49,10 +51,10 @@ class UChaosVDEditorSettings : public UObject
 public:
 
 	UPROPERTY(EditAnywhere, Category = "Viewport Visualization Flags", meta = (Bitmask, BitmaskEnum = "/Script/ChaosVD.EChaosVDParticleDataVisualizationFlags"))
-	uint8 GlobalParticleDataVisualizationFlags = 0;
+	uint32 GlobalParticleDataVisualizationFlags = 0;
 	
 	UPROPERTY(EditAnywhere, Category = "Viewport Visualization Flags", meta = (Bitmask, BitmaskEnum = "/Script/ChaosVD.EChaosVDCollisionVisualizationFlags"))
-	uint8 GlobalCollisionDataVisualizationFlags = 0;
+	uint32 GlobalCollisionDataVisualizationFlags = 0;
 
 	UPROPERTY(EditAnywhere, Category = "Viewport Visualization Flags", meta = (Bitmask, BitmaskEnum = "/Script/ChaosVD.EChaosVDCollisionVisualizationFlags"))
 	bool bShowDebugText = false;
